backend/requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.1
APScheduler==3.10.4
alembic==1.13.3
pytest==8.3.3
pytest-asyncio==0.24.0

